+++ app/src/main/java/agu/analys/viewmodel/RealTradeCoordinator.kt
@@ -95,7 +95,7 @@ class RealTradeCoordinator(
     // Executor delegation
     fun executeCancelRealOrder(symbol: String, orderId: String, onResult: (Boolean, String) -> Unit) =
         executor.executeCancelOrder(symbol, orderId, onResult)
-    fun executeRealTrade(p: String, t: String, pr: Long, a: Double, tp1: Double, tp2: Double, cb: (Boolean, String) -> Unit) =
+    fun executeRealTrade(p: String, t: String, pr: Double, a: Double, tp1: Double, tp2: Double, cb: (Boolean, String) -> Unit) =
         executor.executeTrade(p, t, pr, a, tp1, tp2, cb)
     fun executeRealSellOrders(
         pair: String,
